% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasurementPrecision.R
\docType{package}
\name{MeasurementPrecision-package}
\alias{MeasurementPrecision}
\alias{MeasurementPrecision-package}
\title{MeasurementPrecision: Measurement Precision Toolkit}
\description{
Measurement Precision toolkit for R. This repository contains source code and a package suitable for R 3.0 and above. It computes a few basic statistics and round them accordind to the  measurement instrument (deltax). See Cousineau, D. (submitted) How many decimals? Rounding descriptive statistics based on measurement precision for the formal mathematical derivations of the results.
}
\details{
MeasurementPrecision is a toolkit that computes some basic descriptive 
statistics but also round them based on the measurement precision deltax.
The functions are 

roundMP.fct(fromData = x, deltax, ...)  to proceed from data (vector, matrix, or list of vectors)

roundMP.fct(fromStatistics = list(...), deltax, ...) to proceed from descriptive statistics.

where fct is a summary statistic name.

See help(roundMP) for a list of all the functions.
}
\keyword{internal}
