% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundMP.R
\name{roundMP}
\alias{roundMP}
\title{Measurement Precision Toolkit}
\usage{
roundMP         (x, fct, deltax, ...)
roundMP.mean    (x, deltax, ...)
roundMP.sd      (x, deltax, ...)
roundMP.var     (x, deltax, ...)
roundMP.semean  (x, deltax, ...)
roundMP.ci      (x, deltax, gamma, ...)
roundMP.cohen.d (x, deltax, mu0, ...)
roundMP.t.test  (x, deltax, mu0, ...)
roundMP.meandiff(x, y, deltax, ...)
roundMP.cohen.d (x, y, deltax, ...)
roundMP.t.test  (x, y, deltax, ...)
roundMP.sdpool  (x, y, ..., deltax, ...)
roundMP.F.ratio (x, y, ..., deltax, ...)
}
\arguments{
\item{x}{a vector of numbers;}

\item{fct}{the summary statistic function between quotes: mean, sd,
semean, cohen.d, ci, var, t.test, meandiff, F.ratio, sdpool;}

\item{deltax}{the precision of the instrument;}

\item{y}{(optional) a second vector for bivariate statistics;}

\item{assumptions}{boolean (TRUE to assume relevant symplifying assumptions);}

\item{verbose}{boolean (TRUE to display a human-readable output);}

\item{...}{(optional) any number of vectors for multivariate statistics;}

\item{gamma}{the coverage level for confidence intervals (default if 
omitted 95\%);}

\item{mu0}{for the one-sample cohen.d and one-sample t.test, provide 
the mean of reference;}
}
\value{
the summary statistic and its value rounded based 
                   on the measurement precision
}
\description{
The measurement precision toolkit returns the value of descritive statistics
rounded according to the measurement precision. If measurements are performed 
with a certain precision, called delta_x, then the statistics derived from  
those measurements cannot have more than a certain precisions, computed 
according to the formulas underlying those statistics. 
The descriptive statistics for which an expression of the precision is known
are: 
For univariate statistics: mean, sd (standard deviation),
semean (standard error of the mean), ci (confidence interval),
cohen.d (one-sample Cohen's d, d_1), var (variance), and 
t.test (one-sample t-test);
For bivariate statistics: cohen.d (two-sample Cohen's d, d_p), meandiff 
(mean difference), and t.test (two-sample t-test);
For multivariables: sdpool (pooled standard deviation) and F.ratio 
(only worst-case scenario).
Four scenarios are considered: 
- Extrinsinc precision: precision is estimated
according to a population point of view (uses standard error of the statistic);
- Intrinsinc precision (worst-case): precision is estimated assuming
systematic measurement errors and the maximal impact it can have on the statistic;
- Intrinsinc precision (best-case): precision is estimated assuming non-
systematic measurement errors and the root-mean-squared impact in can have;
- a Middle-ground precision is provided which assumes quasi-systematic
measurement errors weighting equally systematic and non-systematic scenarios.
}
\details{
These functions returns a summary statistic which is rounded
according to the measurement's precision.
}
\examples{
# define a vector (it could be a 1-colum matrix or a one-column data.frame)
x1 <- c(3,4,5)

# get the rounded mean assuming that the instrument is precise to +or- 1
roundMP(x1, 'mean', deltax = 1)
roundMP.mean(x1, deltax = 1)
roundMP.mean(x1, deltax = 1, verbose = TRUE)

# get the rounded standard error, the rounded confidence intervals
roundMP.semean(x1, deltax = 1)
roundMP.ci(x1, deltax = 1)

# get the rounded mean difference between two vectors;
x2 <- c(5,7,9)
roundMP.meandiff(x1, y = x2, deltax = 1)

# get the rounded F ratio, the rounded Cohen's d, and the rounded t-test
# for the last, do not assume symplifiying assumptions
roundMP.F.ratio(x1, y = x2, deltax = 1)
roundMP.cohen.d(x1, y = x2, deltax = 1)
roundMP.t.test( x1, y = x2, deltax = 1, assumptions = FALSE)

# The F ratio and the pooled standard deviation take any number of columns
x3 <- c(2,5,9,11,25)
roundMP.sdpool( x1, x2, x3, deltax = 1)
roundMP.F.ratio(x1, x2, x3, deltax = 1)

}
\references{
\url{https://.../...}
}
\author{
Denis Cousineau, \email{denis.cousineau@uottawa.ca}
}
\keyword{Measurement}
\keyword{precision;}
\keyword{rounding}
